requirements.txt
requests
tabulate
mysql-connector-python